Refactor import paths for FAQ, Hero, and Sponsor sections; add framer-motion and react-icons dependencies
