+++ lib/TextData.ts
