+++ package.json
@@ -48,6 +48,7 @@
         "embla-carousel-react": "^8.3.0",
         "eslint": "8.49.0",
         "eslint-config-next": "13.5.1",
+        "framer-motion": "^12.23.11",
         "input-otp": "^1.2.4",
         "lucide-react": "^0.446.0",
         "next": "^15.3.4",
@@ -57,6 +58,7 @@
         "react-day-picker": "^8.10.1",
         "react-dom": "18.2.0",
         "react-hook-form": "^7.53.0",
+        "react-icons": "^5.5.0",
         "react-resizable-panels": "^2.1.3",
         "recharts": "^2.12.7",
         "sonner": "^1.5.0",
